requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-cors-headers==3.10.0
django-cors-middleware==1.5.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
idna==3.3
Pillow==8.4.0
pkg_resources==0.0.0
psycopg2==2.9.1
pytz==2021.3
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
